Update Histogram to use BigInteger

+++ src/test/java/randomNumbers/Basket.java
@@ -1,13 +1,15 @@
 package randomNumbers;
 
+import java.math.*;
+
 public class Basket {
 
-	public final int upper;
+	public final BigInteger upper;
 	public final String name;
 
 	public int count = 0;
 
-	public Basket(int upper, String name) {
+	public Basket(BigInteger upper, String name) {
 		this.upper = upper;
 		this.name = name;
 	}
